Update private method visibility
It should be possible to extend this driver to modify the behaviour as
required without having to rewrite substantial parts of it.

For example in our use we create a new WebDriver and WebDriverFactory
which extend this project. We use this to add some minor functionality
specific to our project (mostly transitional as we move away from the
Selenium2 driver).

We can also use this layer to fix bugs before they are accepted
upstream.

For example with issue #5 the current implementation of `setValue`
includes triggering a change event by sending a Keyboard Tab character
to move to the next character.

Whilst I would argue that this is a behaviour which includes undesirable
consequences the project may not take this view. I am then left with the
decision of whether to work around this behaviour or to fork the Driver
and use what I deem to be the more correct approach.

If I wish to work around this behaviour in my extension of the driver I
must copy and fix the `setValue()` function, as well as `findElement()`,
`clickOnElement()`, `getDateTimeFormatForRemoteDriver()`, and
`executeJsOnElement()`.

Whilst you could argue that there should be no need to extend the WebDriver
class I would counter that it is often helpful to be able to do so in
certain situations.