Allow reopened closed windows to be switched to
The workaround to support firefox window switching means that the list
of windows can become stale.

When a window is opened, closed, and then reopened the list of windows
then contains two windows with the same name, with the currently open
window likely being later in the list, for example:

    Array
    (
        [4294967304] => questionpreview
        [4294967307] => questionpreview
    )

The switchToWindow code only looks for the first match, which is the
stale window which has since been closed.

Ideally there should be no state stored here, but given the geckodriver
bug that this workaround addresses, the state must be updated more
aggrerssively.