clickOnElement: Ensure element is in view
This code is to work around a bug in geckodriver where the pointer
location is not set before clicking. However, the `moveToElement()`
function only works where the element is in the viewport.

The element must be scrolled into view if it is not already visible
before setting the pointer position.

This issue should also be raised against geckodriver where the bug
actually lies.

Fixes #9